Dev rakib

+++ ClientApp/src/app/views/transferPosting/trainsfer-posting-list/trainsfer-posting-list.component.ts
@@ -11,6 +11,9 @@ import { TransferPostingInfoComponent } from '../transfer-posting-info/transfer-
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
 import { ConfirmService } from 'src/app/core/service/confirm.service';
 import { PaginatorModel } from 'src/app/core/models/paginator-model';
+import { FeaturePermission } from '../../featureManagement/model/feature-permission';
+import { RoleFeatureService } from '../../featureManagement/service/role-feature.service';
+import { AuthService } from '../../../core/service/auth.service';
 
 @Component({
   selector: 'app-trainsfer-posting-list',
@@ -39,6 +42,8 @@ export class TrainsferPostingListComponent implements OnInit, OnDestroy {
   matSort!: MatSort;
   pagination: PaginatorModel = new PaginatorModel();
   noticeForEntryId: number = 0;
+  featurePermission : FeaturePermission = new FeaturePermission;
+  loginEmpId: number = 0;
   
   constructor(
     private toastr: ToastrService,
@@ -47,6 +52,8 @@ export class TrainsferPostingListComponent implements OnInit, OnDestroy {
     private modalService: BsModalService,
     private confirmService: ConfirmService,
     private router: Router,
+    public roleFeatureService: RoleFeatureService,
+    private authService: AuthService,
   ) {
 
   }
@@ -55,11 +62,26 @@ export class TrainsferPostingListComponent implements OnInit, OnDestroy {
 
 
   ngOnInit(): void {
-    this.route.queryParams.subscribe((params) => {
-      this.noticeForEntryId = params['forNotificationId'] || 0;
-      this.getAllTransferPostingInfo(this.pagination);
-    });
-    this.getAllTransferPostingInfo(this.pagination);
+    this.getPermission();
+  }
+
+  getPermission(){
+    this.subscription.push(
+    this.roleFeatureService.getFeaturePermission('transferPostingList').subscribe((item) => {
+      this.featurePermission = item;
+      if(item.viewStatus == true){
+        this.loginEmpId = this.authService.userInformation.empId || 0;
+        this.route.queryParams.subscribe((params) => {
+          this.noticeForEntryId = params['forNotificationId'] || 0;
+          this.getAllTransferPostingInfo(this.pagination);
+        });
+      }
+      else{
+        this.roleFeatureService.unauthorizeAccress();
+        this.router.navigate(['/dashboard']);
+      }
+    })
+    )
   }
 
   getAllTransferPostingInfo(queryParams: any) {
@@ -94,6 +116,15 @@ export class TrainsferPostingListComponent implements OnInit, OnDestroy {
     }
   }
 
+  addTrasnderPosting(){
+    if(this.featurePermission.add == true){
+      this.router.navigate(['/transferPosting/transferPostingApplication']);
+    }
+    else {
+      this.roleFeatureService.unauthorizeAccress();
+    }
+  }
+
   transferPostingInfo(id: number) {
     const initialState = {
       id: id
@@ -101,33 +132,46 @@ export class TrainsferPostingListComponent implements OnInit, OnDestroy {
     const modalRef: BsModalRef = this.modalService.show(TransferPostingInfoComponent, { initialState, backdrop: 'static' });
   }
 
-  delete(element: any) {
-    this.subscription.push(
-    this.confirmService
-      .confirm('Confirm delete message', 'Are You Sure Delete This  Item')
-      .subscribe((result) => {
-        if (result) {
-          this.empTransferPostingService.deleteEmpTransferPosting(element.id).subscribe(
-            (res) => {
-              const index = this.dataSource.data.indexOf(element);
-              if (index !== -1) {
-                this.dataSource.data.splice(index, 1);
-                this.dataSource = new MatTableDataSource(this.dataSource.data);
-              }
-              this.toastr.success('Delete sucessfully ! ', ` `, {
-                positionClass: 'toast-top-right',
-              });
-            },
-            (err) => {
-              this.toastr.error('Somethig Wrong ! ', ` `, {
-                positionClass: 'toast-top-right',
-              });
+  updateInformation(id: number, applicationById: number){
+    if(this.featurePermission.update == true || (this.featurePermission.update == false && applicationById == this.loginEmpId)){
+      this.router.navigate([`transferPosting/update-transferPostingApplication/${id}`]);
+    }
+    else {
+      this.roleFeatureService.unauthorizeAccress();
+    }
+  }
+
+  delete(element: any, applicationById: number) {
+    if(this.featurePermission.delete == true || (this.featurePermission.delete == false && applicationById == this.loginEmpId)){
+      this.subscription.push(
+        this.confirmService
+          .confirm('Confirm delete message', 'Are You Sure Delete This  Item')
+          .subscribe((result) => {
+            if (result) {
+              this.empTransferPostingService.deleteEmpTransferPosting(element.id).subscribe(
+                (res) => {
+                  const index = this.dataSource.data.indexOf(element);
+                  if (index !== -1) {
+                    this.dataSource.data.splice(index, 1);
+                    this.dataSource = new MatTableDataSource(this.dataSource.data);
+                  }
+                  this.toastr.success('Delete sucessfully ! ', ` `, {
+                    positionClass: 'toast-top-right',
+                  });
+                },
+                (err) => {
+                  this.toastr.error('Somethig Wrong ! ', ` `, {
+                    positionClass: 'toast-top-right',
+                  });
+                }
+              );
             }
-          );
-        }
-      })
-    )
-    
+          })
+        )
+    }    
+    else {
+      this.roleFeatureService.unauthorizeAccress();
+    }
   }
